+++ src/components/10-atoms/input-text/index.js
@@ -61,7 +61,7 @@ class AXAInputText extends AXAPopupMixin(NoShadowDOM) {
       onChange: { type: Function, attribute: false },
       onFocus: { type: Function, attribute: false },
       onBlur: { type: Function, attribute: false },
-      onInputfieldKeyUp: { type: Function, attribute: false },
+      onKeyUp: { type: Function, attribute: false },
     };
   }
 
@@ -378,7 +378,7 @@ class AXAInputText extends AXAPopupMixin(NoShadowDOM) {
                   @input="${this.handleInput}"
                   @focus="${this.handleFocus}"
                   @blur="${this.handleBlur}"
-                  @keyup="${this.onInputfieldKeyUp}"
+                  @keyup="${this.onKeyUp}"
                 />
               `
             : html`
@@ -398,7 +398,7 @@ class AXAInputText extends AXAPopupMixin(NoShadowDOM) {
                   @input="${this.handleInput}"
                   @focus="${this.handleFocus}"
                   @blur="${this.handleBlur}"
-                  @keyup="${this.onInputfieldKeyUp}"
+                  @keyup="${this.onKeyUp}"
                 />
               `
         }