+++ src/components/10-atoms/input-text/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 4.5.0
 
-- New callback `onInputfieldKeyUp` which is called if the user presses a key in the input-text field.
+- New callback `onKeyUp` which is called if the user presses a key in the input-text field.
   Can for example be used to recognize when the ENTER key has been pressed.
 
 ## 4.4.0